Confidence and Prediction Intervals using R
Confidence Intervals and Prediction Intervals using R

1. Exercise
a. For the `prostate` data in`faraway`, fit a model with `lpsa` as the response and the other variables as predictors.
b. Compute 90% and 95% CIs for the parameter associated with `age`

2. Exercise
Compute and display a 95% joint confidence region for the parameters associated with `age` and `lbph`. Plot the origin and report the outcome of the appropriate hypotheses test. Affirm this conclusion with an appropriate partial F-test.

3. Exercise
Predict `lpsa` (95%) for a new patient with `lcavol` = 1.22692, `lweight` = 3.62301, `age` = 65, `lbph` = -0.3001, `svi` = 0.0, `lcp` = -0.79851, `gleason` = 7.0, `pgg45` = 15.0. Do this again for the mean response. Using the `exp()` function, obtain the  new prediction and mean response for `psa`.

4. Exercise
Repeat the above exercise with new patient `age` = 20

5. Exercise
For the model in exercise 1, remove all the predictors that are not significant at the 5% level.

6. Exercise
Test the "small" model in exercise 5 against the "big""  model in exercise 1 at *probability type I error* $\alpha = 0.05$. Which model is preferred?