Hypothesis Testing using R
Hypothesis Testing using R

1. Exercise
Run a model with the region as predictor of wages. Show that the number of coefficients associated with the region is 3.

2. Exercise
Apply the `aggregate(wage ~ region, data = uswages, mean)` function in R to obtain the mean `wages` by `region`

3. Exercise
- Compare the two models:
    + Model 1: wage ~ region
    + Model 2: wage ~ region + educ + exper
What is the conclusion - Model 1 or Model 2 is better? So do education and experience matter?

4. Exercise
Compare the two models:
    + Model 1: wage ~ educ + exper
    + Model 2: wage ~ region + educ + exper 

5. Exercise
Repeat exercise #4 using log(wage) for the outcome variable.
Compare the two models:
    + Model 1: log(wage) ~ educ + exper
    + Model 2: log(wage) ~ region + educ + exper