Remedies for Error Problems using R
Exercise 1
a. Create factors, remove NAs
b. Process the data to make factors with named levels for *race*, *smsa*, and *pt*.
c. Replace the indicator variables *ne*, *mw*, *so*, and *we* with one factor called *region* with levels named: ne*, *mw*, *so*, and *we*.
d. Remove all NAs.

# Exercise 2
a. Show that the summary of your final data is the following.

# Exercise 3
a. Compute OLS fit to model log(wage)~.
b. Perform the Shapiro-Wilk Test of Normality for the residuals, what is the conslusion?

 Exercise 4
a. Compute WLS fit to model log(wage)~. and weights = 1/(1+educ) 
b. Perform the Shapiro-Wilk Test of Normality for the residuals, what is the conslusion?

# Exercise 5
a. Compute Robust fit to model log(wage)~. using Huber, Hampel, Biquare, LTS, and LAD
b. Compare coefficients of the above fits using OLS, WLS, Huber, Hampel, Biquare, LTS, and LAD
c. Which would you recommend?
d. Why?